move tests to model bom file
