+++ tests/test_component.py
@@ -24,9 +24,8 @@ from unittest import TestCase
 from packageurl import PackageURL  # type: ignore
 
 from cyclonedx.model import sha1sum
-from cyclonedx.model.bom import Bom
 from cyclonedx.model.component import Component
-from data import get_component_setuptools_simple, get_component_setuptools_simple_no_version
+from data import get_component_setuptools_simple
 
 FIXTURES_DIRECTORY = 'fixtures/xml/1.4'
 
@@ -78,20 +77,3 @@ class TestComponent(TestCase):
         )
         self.assertEqual(c.purl, purl)
         self.assertEqual(len(c.hashes), 1)
-
-    def test_has_component_1(self) -> None:
-        bom = Bom()
-        bom.components.update([get_component_setuptools_simple(), get_component_setuptools_simple_no_version()])
-        self.assertEqual(len(bom.components), 2)
-        self.assertTrue(bom.has_component(component=get_component_setuptools_simple_no_version()))
-        self.assertIsNot(get_component_setuptools_simple(), get_component_setuptools_simple_no_version())
-
-    def test_get_component_by_purl(self) -> None:
-        bom = Bom()
-        setuptools_simple = get_component_setuptools_simple()
-        bom.components.add(get_component_setuptools_simple())
-
-        result = bom.get_component_by_purl(get_component_setuptools_simple().purl)
-
-        self.assertEqual(result, setuptools_simple)
-        self.assertIsNone(bom.get_component_by_purl(get_component_setuptools_simple_no_version().purl))
